// frontend/src/app/models/Dostava.ts
export class Dostava {
  _id: string = '';
  id: number = 0;
  gost: string = '';
  restoran: number = 0;
  ime_prezime: string = '';
  adresa_gosta: string = '';
  naziv_restorana: string = '';
  datum: number = 0;
  sadrzaj: string[] = [];
  racun: number = 0;
  status: string = '';
  procenjeno_vreme: number = 0;
}